领导管辖部门

+++ src/main/java/edu/cqu/ncycoa/plan/PlanHelper.java
@@ -1,11 +1,18 @@
 package edu.cqu.ncycoa.plan;
 
+import java.util.ArrayList;
 import java.util.Calendar;
+import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 import java.util.TreeMap;
 
 import com.db.DBObject;
+import com.db.DataRow;
+import com.db.DataTable;
 import com.db.Parameter;
+import com.entity.system.Org;
+import com.entity.system.StaffInfo;
 import com.performance.ParaDataHelper;
 
 public class PlanHelper {
@@ -72,8 +79,8 @@ public class PlanHelper {
 			DBObject db=new DBObject();
 			String value = db.executeOneValue("select count(*) from tbm_admindpt where staffcode='" + staffcode + "'");
 			if("0".equals(value)){
-				String sql="insert into tbm_admindpt(staffcode) values(?)";
-				Parameter.SqlParameter[] paras = new Parameter.SqlParameter[]{new Parameter.String(staffcode)};
+				String sql="insert into tbm_admindpt(staffcode, adminmode) values(?, ?)";
+				Parameter.SqlParameter[] paras = new Parameter.SqlParameter[]{new Parameter.String(staffcode), new Parameter.String("工作")};
 				db.run(sql, paras);
 			}
 		} catch (Exception e) {
@@ -84,10 +91,10 @@ public class PlanHelper {
 	public static void insertTask(String staffcode, String orgcode){
 		try {
 			DBObject db=new DBObject();
-			db.run("delete from tbm_admindpt where staffcode='" + staffcode + "' and (orgcode is null or orgcode='"+orgcode+"'");
+			db.run("delete from tbm_admindpt where staffcode='" + staffcode + "' and adminmode='工作' and (orgcode is null or orgcode='"+orgcode+"')");
 			
-			String sql="insert into tbm_admindpt(staffcode,orgcode) values(?,?,)";
-			Parameter.SqlParameter[] paras = new Parameter.SqlParameter[]{new Parameter.String(staffcode), new Parameter.String(orgcode)};
+			String sql="insert into tbm_admindpt(staffcode,orgcode,adminmode) values(?,?,?)";
+			Parameter.SqlParameter[] paras = new Parameter.SqlParameter[]{new Parameter.String(staffcode), new Parameter.String(orgcode), new Parameter.String("工作")};
 			db.run(sql, paras);
 		} catch (Exception e) {
 			e.printStackTrace();
@@ -97,7 +104,7 @@ public class PlanHelper {
 	public static void delTask(String staffcode){
 		try {
 			DBObject db=new DBObject();
-			db.run("delete from tbm_admindpt where staffcode='" + staffcode + "'");
+			db.run("delete from tbm_admindpt where staffcode='" + staffcode + "' and adminmode='工作'");
 		} catch (Exception e) {
 			e.printStackTrace();
 		}
@@ -106,15 +113,61 @@ public class PlanHelper {
 	public static void delTask(String staffcode, String orgcode){
 		try {
 			DBObject db=new DBObject();
-			db.run("delete from tbm_admindpt where staffcode='" + staffcode + "' and orgcode='" + orgcode + "'");
+			db.run("delete from tbm_admindpt where staffcode='" + staffcode + "' and adminmode='工作' and orgcode='" + orgcode + "'");
 		} catch (Exception e) {
 			e.printStackTrace();
 		}
 	}
 	
 	
+	public static List<StaffInfo> getStaff(){
+		List<StaffInfo> returnValue = new ArrayList<StaffInfo>();
+		Map<String, Integer> flag= new HashMap<String, Integer>();
+		try {
+			DBObject db=new DBObject();
+			
+			String sql="select staffcode from tbm_admindpt where adminmode = '工作'";
+			DataTable dt=db.runSelectQuery(sql);
+			if(dt!=null){
+				for(int i=0;i<dt.getRowsCount();i++){
+					DataRow r=dt.get(i);
+					
+					String staffcode = r.getString("staffcode");
+					if(flag.get(staffcode) == null){
+						flag.put(staffcode, 1);
+						returnValue.add(new StaffInfo(staffcode));
+					}
+				}
+			}
+		} catch (Exception e) {
+			e.printStackTrace();
+		}
+		
+		return returnValue;
+	}
 	
-	
+	public static List<Org> getOrgs(String staffcode){
+		List<Org> returnValue = new ArrayList<Org>();
+		try {
+			DBObject db=new DBObject();
+			
+			String sql="select orgcode from tbm_admindpt where adminmode= '工作' and staffcode='" + staffcode + "'";
+			DataTable dt=db.runSelectQuery(sql);
+			if(dt!=null){
+				for(int i=0;i<dt.getRowsCount();i++){
+					DataRow r=dt.get(i);
+					String tmp = r.getString("orgcode");
+					if(tmp != null && !"".equals(tmp)){
+						returnValue.add(new Org(r.getString("orgcode")));
+					}
+				}
+			}
+		} catch (Exception e) {
+			e.printStackTrace();
+		}
+		
+		return returnValue;
+	}
 	
 //	public static String getDataGrid(ReviewTask task, Map<String, HashMap<String, String>> data) throws ReviewException{
 //		if(task == ReviewTask.INVALID_TASK){
